+++ Changelog.md
@@ -3,6 +3,10 @@
 1. Fix issue #55.
     * [Pull request 175](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/175)
 
+1. Do not allow to advertise a topic that is currently advertised on the same node.
+   See [issue #54] (https://bitbucket.org/ignitionrobotics/ign-transport/issues/54)
+    * [Pull request 180](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/180)
+
 ### Ignition Transport 1.2.0
 
 1. Removed duplicate code in NetUtils, and improved speed of DNS lookup